requirements.txt
playwright>=1.40.0
requests>=2.31.0
pyyaml>=6.0.1
setuptools>=65.0.0
